Moved database-logic to seperate file
